ComposeSceneRecomposer
